organized files into a use case structure
