+++ frontend/src/App.tsx
@@ -1,17 +1,17 @@
-// import "./App.css";
+import "./App.css";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import { Desktop } from "./components/Desktop/Desktop";
-import MainPage from "./components/MainPage";
-import UploadDataset from "./components/UploadDataset";
-import GuidancePage from "./components/GuidancePage/GuidancePage";
+import UploadDataset from "./pages/UploadDataset";
+import HomePage from "./pages/HomePage/HomePage";
+import GuidancePage from "./pages/GuidancePage/GuidancePage";
+import DashboardPage from "./pages/DashboardPage/DashboardPage";
 
 function App() {
   return (
     <Router>
       <Routes>
-        <Route path="/" element={<Desktop />} />
+        <Route path="/" element={<HomePage />} />
         <Route path="/upload-dataset" element={<UploadDataset />} />
-        <Route path="/graph" element={<MainPage />} />
+        <Route path="/graph" element={<DashboardPage />} />
         <Route path="/guidance" element={<GuidancePage />} />
       </Routes>
     </Router>
