--- frontend/src/components/Button.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-const Button: React.FC = () => {
-    return (<button style={{ color: 'black' , backgroundColor:'green', padding: '10px 20px'}}>
-        Dashboard Button
-        </button>
-    );
-};
-
-export default Button;
