--- frontend/src/components/Title.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Title: React.FC = () => {
-    return <h1 style={{ color: 'green' }}>Breaking Bias</h1>;
-};
-
-export default Title;
