+++ frontend/src/pages/DashboardPage/components/ExportGraphButton.tsx
@@ -1,5 +1,5 @@
 import { Button } from "@mui/material";
-import { ExportService } from "../services/ExportService";
+import { ExportService } from "../../../services/ExportService";
 
 function ExportGraphButton() {
   function handleExport() {
