+++ frontend/src/pages/DashboardPage/components/Slider.tsx
