+++ frontend/src/pages/GuidancePage/GuidancePage.css
