+++ frontend/src/pages/GuidancePage/GuidancePage.tsx
